Added factory_girl for test object generation, along with tests to test the testing system.
factory_girl (and factory_girl_rails) from:
https://github.com/thoughtbot/factory_girl
https://github.com/thoughtbot/factory_girl_rails

Factory_girl has some built-in steps for Cucumber. For each factory, there is:
* Given the following factory_name(s) exists -> |table|
* Given a factory_name exists
* Given # factory_names exist
And for each column in an ActiveRecord class factory:
* Given a factory_name exists with a column_name of "value"
* Given # factory_names exist with a column_name of "value"

Added a TestModel with a testing_system feature and a test_model_spec to verify the functioning of the testing systems (Rspec, Cucumber, factory_girl) in my configuration.