More features defined for user access.
These are just the cucumber feature definitions (flagged as @future for now, so they don’t actually run).

Also copied over a bunch of cucumber step definitions (taken mostly from those that come with clearance) to support the features. If you take away the @future tag from the features and run cucumber, the tests will fail, but there won’t be any undefined steps.

Also adjusted cucumber date-setting step to allow the optional word “now”. So, you can now do:
> Given the date is "some date"
or
> Given the date is now "some other date"
A minor detail, but it reads better in some contexts.

Also added chronic gem for development/testing as it is relied on by the date setting steps in cucumber.