Cucumber feature testing of user access.
Added user access menu to top of application layout. Note that the format of the ul element is important in that it is used by cucumber testing to determine whether the user is signed in or not.

Really ugly monkey-patch in the user steps in order to simulate a failure condition.