Authority model — limiting access to items.
The Authority model can be applied to any ActiveRecord model via the ‘acts_as_authority_controlled’ method (found in ‘lib/authority_controlled.rb’). Instructions for the various configuration options are in the source and ‘doc/Developers/Authorities.rdoc’.

The initial framework for the model was generated with:
> rails g scaffold Authority user:belongs_to item:belongs_to area:string is_owner:boolean can_create:boolean can_view:boolean can_edit:boolean can_delete:boolean can_invite:boolean can_permit:boolean

Note that I haven’t committed the controller & views files because I haven’t done anything with those yet.