Minor tweaks to cucumber web steps.
I escape the quote in regexps so that my syntax colouring works nicely. Doesn’t affect the actual code.

The “When (.*) within ([^:]+)” step was not being used, but it was interfering with another step, so I commented it out.

I changed the “Then I should be on (.*)” step to allow for regular expressions checking instead of straight string equality.