AuthoritiesController implemented.
Also a pile of changes and additions to the:
• Authority and User models.
• cucumber step definitions for users and authorities.
• user and authority feature scenarios.
• authority_controlled library.
• user-related specs.

Note that I’m caving in to the pressure to switch from using tab indentation to the Ruby community’s 2-space indentation. So, some diffs between files will have a bunch of “extra” non-matching lines where only the indentation has changed.