Page version tracking, part 1.
Pages create a new Version record whenever they are saved. The page.editor must be set to a valid user for every save.

Versions are never directly created, but instead are created when a versioned record (such as a Page) is saved.
Versions are never updated. If an update is needed, it should be a new version.

The versions controller should always be a sub-controller. Currently, it’s just used as a sub-controller for pages (“/pages/123/versions”).

Updated authority_controlled to handle the per-item restriction of Pages.

Still to do:
• Integrate version access into page display.
• Nicer version display and navigation.