Gem updates (including Rails 3.0.8); switch from Webrat to Capybara for simulated browser testing.
Some tweaking of Cucumber scenarios and step definitions, primarily to accommodate the switch to Capybara. Note that the Cucumber testing of OAuth functions is a little wonky because of Capybara not happily retaining session state when hitting remote urls.

Miscellaneous indentation cleanup (switch to two-spaces instead of tab characters). There are still legacy files using tabs for indentation. The gradual conversion proceeds as files get touched.