Filled out missing specs for Paths Controller (100% coverage).
Reimplemented the authority restrictions on paths based mostly on the Pages Controller.

Changed authority_controlled library so that models that inherit authorities (such as how paths inherit from their items) work when they don’t have an item to inherit from (such as when paths are just redirects instead of being attached to an item).