Authority controls for Documents.
Documents can now be individually set to require authority to access. At least, programmatically. There’s still no interface for actually adding authorities to a document.

I’ve learned about RSpec `context` now, so have started to use that to make the spec tests clearer, hopefully.

Pages auto-create their first `version` when saved, so changed the versions factory to require passing in an item, so as to avoid ending up with multiple versions on an item when not intended.