Gem updates. Bye-bye MetricFu.
Because of gem conflicts and other annoyances, I’ve dropped MetricFu, for now. I may try to re-add some of it’s components.

Got a little more restrictive on some of the gem versions (using `~>` rather than `>=` in some cases).

Also regrouped many of the gems that were in the `:development` & `:test` group to just the `:test` group.