[JavaScript] character counter function.
Used to display the number of characters left that can be entered in a text field/area.

Add `onkeyup="countChars(...)"` to the `input` or `textarea` element.
* char_limit: an integer of the character limit for the field.
* field_id: the `id` attribute for the form field whose characters are being counted.
* count_display_id: the `id` attribute of the html element containing the count (typically a `span`).

Add an element, such as a `span`, which will be updated to contain the limit count (it’s contents will be overwritten by the function call).