[Event] added is_cancelled and timezone attributes.
Allow an event to be flagged as cancelled (kept in the system, and shown in normal listings, but identified as cancelled).

Track the timezone an event is in (if applicable).

Use better form fields (url_field, custom datetime_field).

Also reordered validations on `Event` to match the order the fields appear in in the forms.

Some simplification of the events model and controller specs.