[Event][Version] versioning of events.
Use the Version model to track changes to Event records. Auto-generated whenever an Event is saved.

Moved accessing of current_user to a before_filter on EventsController (d.r.y.).

Using the version count as the `SEQUENCE` number for icalendar representations of events.

Added versions link (when more than one version of an item) when displaying an item using the application html layout.

Using factories in EventsController spec to reduce complexity (otherwise, would have to go in and assign an editor every time).