Constants for site name and description
`Wayground::Application::NAME` and `Wayground::Application::DESCRIPTION` are used in various views instead of hard-coded strings.

Added `X-WR-CALNAME` and `X-WR-CALDESC` to the icalendar (.ics) layout so that people subscribing to a calendar will get something useful in the calendar name and description.

Also set the default home view to tell browsers & robots not to cache the page, since it hasn’t been created yet if they’re getting the default home template.