[Gem] update gems, especially FactoryGirl and Rspec.
The update to FactoryGirl deprecated the older version’s syntax, so affected many of the spec & feature files. However, the code impact was pretty simple — just a change to the method calls, not the parameters or functions. There was substantive impact on structure for the factory definitions themselves (`spec/factories.rb`).

There's a bug (or maybe it was intentional — in which case: ugh) in Rspec where it now goes to the default view template format (html) instead of the one referenced in the view spec file name or `describe` name. So, at least for now, we have to explicitly give the template path when calling render in view specs for view templates using non-html formats (such as `ics` formats like `spec/views/events/show.ics.erb_spec.rb`).