[Gem] Update to Rails 3.1.4.
Had to slightly revise how files are passed for upload in controller testing. This lead to a minor change to the `Document#file=` method.

Moved the Javascript and CSS assets to fit Rails new asset hierarchy.

The `active_support/core_ext/string/output_safety.rb` method `html_escape` (in the activesupport-3.1.4 gem) spits out a warning (that doesn’t break functionality) when passed a non-html_safe UTF string. I’m not addressing that here because it gets fixed in Rails 3.2.