[IcalProcessor] replace icalendar gem with a simpler reader lib.
The icalendar gem was proving too convoluted for the simple functionality I actually wanted to get out of it. So, I wrote IcalendarReader (`lib/icalendar_reader.rb`) as a simple iCalendar parser. It was mostly very straight-forward, but about 80% of my time on it was just dealing with parsing dates because of the pain that is timezones and daylight savings time. Grrrrr.

It doesn’t handle things like VALARM or recurrence rules because I’m not supporting alarms or recurrence in the Event model, currently.

Added the `tzinfo` gem to help interpret timezones.