[Event] set locally modified flag on sourced_items when updating other than from source.
An `#is_sourcing` flag (temporary, not saved in the record) gets set on an Event when it’s being created/updated from a Source.

When not sourcing, any sourced_items on the Event have their `#has_local_modifications` field set to true when the Event is updated.

Added a sourced_item factory for testing.