[CSS][Layout] convert to SASS (.scss) and overhauled design.
Beginning to use SASS for stylesheets. A lot of the existing CSS could still benefit from being updated to take advantage of SASS tools.

Did a significant reworking of the CSS and the application layout view. Now a bit more HTML5-ish, as well as flowing and adapting better between tight mobile screens and wide screens.