[CSS][Layout] redid the navigation bars.
Added a bunch of mixins to the SASS helpers.

`@page_submenu_items` can be set to an array of hashes `[{title: ?, path: ?},…]` that will be used to add items to the page submenu (which currently appears as a fixed bar along the bottom of the page).

Changed `@site_section` from a string to a symbol (e.g., `'Events'` is now `:events`).

Dropped `@page_uses_javascript` because Javascript is now part of the core layout, so used on every page.