[Gem] Update to Rails 3.2.6.
Updated confirm attributes on links in views to use the new rails syntax. (`:confirm => 'text'` changed to `:data => {:confirm => 'text'}`.)

Also fixed a bug (in the events show view) that hadn’t previously turned up where the lack of a location url on an Event would cause an error if the location city was set but the location and location address were blank.