[Gem] Brakeman gem for security analysis.
Checks rails apps for common security risks.
http://brakemanscanner.org/

Call from a rake task in the root of the project:
> bundle exec rake brakeman
That will generate a `brakeman.html` report file.

Fixed a couple validation problems brakeman identified.

Force the redirect urls handled by the oauth callback in the sessions controller to be local paths.