[Event][IcalProcessor] Move some ical methods from Event to IcalProcessor, and some test cleanup.
New method `event#perform_from_source` is a wrapper that takes a block and runs the block wrapped in a flag that causes operations to act as if the Event is coming from a Source (to avoid certain aspects of callbacks that operate on Sourced Events).

Changed methods with multiple arguments to take a hash instead.