[UserToken][User][Rememberer] refactor `User#remember_token` to a separate class.
Responsibility for the logic of remembering users is moved to the new `Wayground::Rememberer` model.

Responsibility for persistence of tokens to remember Users is now in the `UserToken` ActiveRecord model.

Setting of cookies in controllers has now moved to methods in the application_controller (`cookie_set_remember_me` and `cookie_set_remember_me_permanent`).