[OauthLogin][Authentication] refactor OAuth authentication into Login::OauthLogin class.
Moved the OAuth authentication logic from the `Authentication` ActiveRecord model into the `OauthLogin` class. (Separating persistence from logic.)

Changed behaviour when a user is logged in and authenticates with an existing authentication for a different user. It now changes the login to the different user.

Some cleanup of User model, `user_spec.rb`, and `sessions_controller_spec.rb`.