[EventPresenter] html presenter for event data.
I built this is for use by the impending calendar-format views, but I’ve also changed the `app/views/events/_events.html.erb` partial template to use this presenter now.

I created a ViewDouble class for convenience in testing, but I’m sure there’s a better way to take care of that. I recall seeing a method for Rspec to give a spec a view context, but I’ve been unable to find that again.