[CalendarController][CalendarMonthPresenter] view events in calendar format.
CalendarController provides views by year, month or day.

The `CalendarMonthPresenter` is an html presenter for generating the grid of weeks to display a month in a calendar.

Added a helper `calendar_day_path_for_date` for getting the path to a calendar day by passing in a date (instead of having to parse out the date elements and ensure they are zero-padded strings).