[!!!][PostgreSQL] database switched from sqlite3 to PostgreSQL.
This necessitated a number of minor adjustments in the code.

`ExternalLink.default_scope` set so position will be used to sort.

Dropped the unused `Version.current_versions` scope since it wouldn’t work with Postgres without a bunch of changes to it.

Constrained `Wayground::Rememberer#get_random_token_string` to a max 127 characters to fit in the new default string data size.