[VersionsController] set nocache flag on index and show actions.
Don’t want robots to index the versions — they should just index the main views of the items. (E.g., index an event’s page, not the versions of it.)

Consistently use `:nocache` as the parameter name for metadata (replacing the inconsistent use of `cache: :no`).

Also changed hash notation to ruby 1.9 format, in affected files.