[gem] more minor gem updates.
The `main` gem, which `churn` depends on, apparently forgot to list its dependencies in it’s latest version. So, I’ve had to (probably temporarily) list them in the Gemfile.

Also did
* `gem update --system` (now version 2.2.2)
* `gem update bundler` (now version 1.5.3)
which, because they sit outside the source project, aren’t reflected in the Gemfile or anywhere else in source control.