[Rails] upgrade Rails to 4.1.8.
This is an interim step before upgrading to Rails 4.2. This is continuing the process of doing the last sub-point releases under each point release in sequence. (The previous was the upgrade to 4.0.12 from 4.0.4.)

Changes to code in this commit are just straight copies over from a new, default, Rails 4.1.8 project generated with: `rails new railsversion --database=postgresql --skip-test-unit` (where `railsversion` is just a throwaway project name).

Also had to make a few adjustments around the values stored in `Version` objects. The hash keys are defaulting to strings now, without supporting symbols in some cases. Also, values are being represented as strings. I don’t know why it’s become less flexible about this, but it has had only a small impact (just a couple specs, and the `Version#diff_with` method, seem to be affected).

Also found an intermittent bug (thanks to RSpec’s randomization of test order) in SourcedItem which required a slight modification to `SourcedItem#set_date`.