[Rails] upgrade Rails to 4.2.0.
Changes to code in this commit are just straight copies over from a new, default, Rails 4.2.0 project generated with: `rails new railsversion --database=postgresql --skip-test-unit --skip-turbolinks` (where `railsversion` is just a throwaway project name).

Using a custom-patched version of the `rspec-autotest` gem while waiting for a new version of the real gem to be patched to fix its deprecations bug.