[TagsController] new controller for accessing stuff through tags.
The `index` action displays a “word cloud” of the tags.

The (incomplete) `tag` action is intended to eventually show all items that have the given tag (other than items the user doesn’t have permission to view).

When that `tag` action is working, then the `index` view should be updated to link the tags it displays to the `tag` action.