[SourcesController] `runprocessor` action changed to display affected events.
Now directly rendering the `show` template instead of redirecting. That is so we can assign the list of `sourced_items` to be displayed to be just the new items created and the updated items (not bothering to show the items skipped because of no changes since last update from the source).

Also expanded the testing on the `runprocessor` action.