[CandidatesController] handling of specifying an existing Person record when creating a new Candidate record.
Moved the assignment of the person to after the setting of the params on the new candidate, so values from the person record will be used if not explicitly set in the params.

Added a person field to the form (editable field for `new` action, static text for `edit` action).

Made the name field bigger.