[gem][preflight] Add Rubocop gem, and remove redundant code analysis tools.
A few of the code analysis gems weren’t providing any additional value over what is in Rubocop’s reports.
* cane
* flog
* roodi
* tailor

Adding Rubocop necessitated a slight change in the `.ruby-version` file. The effect of the file remains unchanged (use MRI 2.3.0), but rubocop couldn’t recognize when the `ruby-` prefix was on the version number. So, I removed the prefix.

Also made some adjustments to some of the other gem calls in the `preflight` script.