[gem][spec][setup] Added html_validation gem; feature specs config.
This will be used for testing the validity of the html generated in feature specs.

INSTALLATION:

The html_validation gem requires the `tidy` command-line tool.

See `doc/Developers/Setup.md` for installation tips.

---

There’s some code (embedded in `spec/feature_helper.rb`) that overrides the gem behaviour to allow for setting tidy options when calling the validation test.
This should really be cleaned up and submitted as a proper pull-request on the gem, but I’m being quick‘n’dirty on this at the moment.