[spec][feature] Test validity of front page html.
A first run at using Rspec for feature tests (instead of Cucumber, which I haven’t touched in ages now).

Also the first use of html validation in the testing suite. While validation testing should be done for every page generated in the system, this at least confirms the general application layout as valid, for now.